+++ src/lbm_bnd.h
@@ -4,8 +4,9 @@
 #include <sfcmm_common.h>
 #include "common/surface.h"
 #include "configuration.h"
-#include "lbm_bnd_wall.h"
+#include "lbm_bnd_in_out.h"
 #include "lbm_bnd_interface.h"
+#include "lbm_bnd_wall.h"
 
 template <LBMethodType LBTYPE, GBool TANGENTIALVELO>
 class LBMBnd_wallBB;
@@ -38,21 +39,33 @@ class LBMBndManager : private configuration {
           addBndry(BndryType::Periodic, bndrySurface(surfId), surfBndConfig);
           continue;
         }
-        if(bndType == "wall"){
+        if(bndType == "wall") {
           const GDouble tangentialV = config::opt_config_value(surfBndConfig, "tangentialVelocity", 0.0);
-          if(std::abs(tangentialV) > GDoubleEps){
+          if(std::abs(tangentialV) > GDoubleEps) {
             logger << " wall bounce-back with tangential velocity" << std::endl;
             addBndry(BndryType::Wall_BounceBack_TangentialVelocity, bndrySurface(surfId), surfBndConfig);
-          } else{
+          } else {
             logger << " wall bounce-back with no slip" << std::endl;
             addBndry(BndryType::Wall_BounceBack, bndrySurface(surfId), surfBndConfig);
           }
           continue;
         }
+        if(bndType == "outlet"){
+          logger << " outlet bounce-back with constant pressure" << std::endl;
+          addBndry(BndryType::Outlet_BounceBack_ConstPressure, bndrySurface(surfId), surfBndConfig);
+          continue;
+        }
+        if(bndType == "inlet"){
+          logger << " inlet bounce-back with constant pressure" << std::endl;
+          addBndry(BndryType::Inlet_BounceBack_ConstPressure, bndrySurface(surfId), surfBndConfig);
+          continue;
+        }
+        TERMM(-1, "Invalid bndCndType: " + bndType);
       }
     }
   }
 
+  //todo: merge with the function above
   void addBndry(const BndryType bnd, const Surface<dim(LBTYPE)>& surf, const json& properties) {
     ASSERT(surf.size() > 0, "Invalid surface");
     switch(bnd) {
@@ -62,11 +75,21 @@ class LBMBndManager : private configuration {
       case BndryType::Wall_BounceBack_TangentialVelocity:
         m_bndrys.emplace_back(std::make_unique<LBMBnd_wallBB<LBTYPE, true>>(surf, properties));
         break;
+      case BndryType::Inlet_BounceBack_ConstPressure:
+      case BndryType::Outlet_BounceBack_ConstPressure:
+        m_bndrys.emplace_back(std::make_unique<LBMBnd_InOutBB<LBTYPE>>(surf, properties));
+        break;
       case BndryType::Periodic:
         m_bndrys.emplace_back(std::make_unique<LBMBnd_dummy>());
         break;
       default:
         TERMM(-1, "Invalid bndry Type!");
+    }
+  }
+
+  void preApply(const std::function<GDouble&(GInt, GInt)>& f, const std::function<GDouble&(GInt, GInt)>& fold) {
+    for(const auto& bndry : m_bndrys) {
+      bndry->preApply(f, fold);
     }
   }
 
@@ -88,6 +111,7 @@ class LBMBnd_dummy : public LBMBndInterface {
  public:
   void init() override {}
 
+  void preApply(const std::function<GDouble&(GInt, GInt)>& f, const std::function<GDouble&(GInt, GInt)>& fold) override {}
   void apply(const std::function<GDouble&(GInt, GInt)>& f, const std::function<GDouble&(GInt, GInt)>& fold) override {}
 };
 #endif // LBM_BND_H